Add new document creation
From the menu you can now create a new document, it generates random
document name. The document changes are now also sent to the server, but
not yet pulled from the server. I also realized that I need a method to
save the document. Document saving basically is removing all changes and
saving the current data in one place.